+++ app/notify/page.tsx
@@ -1,19 +1,15 @@
 'use client'
 
 import { NotifyGroup } from '~/components/notify/notify-group'
+import { useNotifyStore } from '~/stores/notify-store'
 import { Tables } from '~/types/supabase'
-import { useNotify } from '~/hooks/useNotify'
 
 type GroupedNotify = {
   [crated_at: string]: Tables<'notifications'>[]
 }
 
 function NotifyPage() {
-  const { notify, loading } = useNotify()
-
-  if (loading) {
-    return <div>로딩 중...</div>
-  }
+  const { notify } = useNotifyStore()
 
   const groupedData = notify.reduce<GroupedNotify>((acc, curr) => {
     const createdAt = curr.created_at?.split('T')[0] || 'Unknown'
@@ -33,7 +29,7 @@ function NotifyPage() {
             <NotifyGroup key={createdAt} createdAt={createdAt} items={items} />
           ))
         ) : (
-          <li>알림이 없습니다.</li>
+          <li>로딩중입니다!</li>
         )}
       </ul>
     </div>
