+++ app/settings/password-and-security/page.tsx
@@ -1,7 +1,6 @@
 'use client'
 
 import { PasswordForm } from '~/components/password-and-security/password-form'
-import { Toaster } from '~/components/ui/sonner'
 
 function PasswordAndSecurity() {
   return (
@@ -16,7 +15,6 @@ function PasswordAndSecurity() {
       <div className="rounded-lg bg-background px-5 py-5">
         <PasswordForm />
       </div>
-      <Toaster />
     </div>
   )
 }
